+++ driver-core/CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 - [improvement] Handle the static columns introduced in Cassandra 2.0.6 (JAVA-278)
 - [bug] Add missing iso8601 patterns for parsing dates (JAVA-279)
 - [bug] Properly parse BytesType as the blob type
+- [bug] Potential NPE when parsing schema of pre-CQL tables of C* 1.2 nodes (JAVA-280)
 
 Merged from 1.0 branch: